Back end refactoring to separate database module
