+++ trackrecord-be/src/database/database.module.ts
@@ -1,0 +1,9 @@
+// from NestJS Auth demo
+import { Module, HttpModule } from '@nestjs/common';
+import { DatabaseService } from './database.service';
+
+@Module({
+  providers: [HttpModule, DatabaseService],
+  exports: [DatabaseService],
+})
+export class DatabaseModule {}
