+++ trackrecord-be/src/users/users.module.ts
@@ -1,10 +1,11 @@
 // from NestJS Auth demo
 import { Module, HttpModule } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { DatabaseService } from './database.service';
+import { DatabaseModule } from '../database/database.module';
 
 @Module({
-  providers: [HttpModule, UsersService, DatabaseService],
+  providers: [HttpModule, UsersService],
   exports: [UsersService],
+  imports: [DatabaseModule]
 })
 export class UsersModule {}
